PatchReadFirstLane: restructure collectAssumeUniforms
The decision of whether an instruction should be added to the
m_canAssumeUniformDivergentUseMap is only made once all later
instructions in the basic block have been processed. To avoid scanning
all instructions excessively, we maintain a basic-block-ordered queue
of candidates. An instruction is only added as a candidate if it appears
as a relevant operand to an instruction that is already in the map.